refactor: centralize cartesian utils
