--- client/src/mixins/cartesian.js
@@ -1,19 +1,0 @@
-export const cartesian = {
-  build(length) {
-    const zeros = Array
-      .apply(null, Array(length))
-      .map(Number.prototype.valueOf, 0);
-
-    return zeros.map(function (i) {
-      return zeros.slice();
-    });
-  },
-  iterate(array, callback) {
-    array.forEach((row, x) => {
-      row.forEach((cell, y) => {
-        callback(Number(x), Number(y))
-      });
-    });
-  },
-};
-export default cartesian;
