+++ shared/utils/cartesian.js
@@ -1,8 +1,6 @@
-const cartesian = {
+export const cartesian = {
   build(length) {
-    const zeros = Array
-      .apply(null, Array(length))
-      .map(Number.prototype.valueOf, 0);
+    const zeros = Array.from({ length }, () => 0);
     return zeros.map(() => zeros.slice());
   },
   iterate(array, callback) {
@@ -11,11 +9,6 @@ const cartesian = {
         callback(Number(x), Number(y));
       });
     });
-  },
-  inBounds(num, size) {
-    if (num < 0) return 0;
-    if (num >= size) return size - 1;
-    return num;
   },
 };
